simple-chat/requirements.txt
Flask==1.1.1
Flask-Sockets==0.2.0
gevent==21.1.2
gevent-websocket==0.10.1
greenlet==1.0.0
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
redis==3.5.3
websocket-client==0.58.0
Werkzeug==1.0.0
six==1.16.0
python-dotenv-0.17.1

